requirements.txt
openai
python-dotenv
colorama
elevenlabslib
halo